memory/requirements.txt
mem0ai>=0.1.0
qdrant-client>=1.7.0
psycopg2-binary>=2.9.0
redis>=4.5.0
httpx>=0.25.0
openai>=1.0.0
anthropic>=0.20.0
cryptography>=41.0.0
